re-enable kill file for graceful exit from uploadServer (possible because ProjectSend uploader prefixes uploaded files with user, timestamp tags)

+++ inst/scripts/killMotusUploadServer.sh
@@ -1,23 +1,23 @@
 #!/bin/bash
 #
 # kill the motus upload server if it is running
-## NOT WORKING YET:  Specify "-g" ("graceful") to let it finish its current subjob first.
+# Specify "-g" ("graceful") to let it finish its current subjob first.
 
-# KILLFILE=/sgm/uploads/kill
-# GRACEFUL=""
-# if [[ "$1" == "-g" ]]; then
-#     GRACEFUL=1
-#     shift
-# fi
+KILLFILE=/sgm/uploads/kill
+GRACEFUL=""
+if [[ "$1" == "-g" ]]; then
+    GRACEFUL=1
+    shift
+fi
 
 PIDFILE=/sgm/uploadServer.pid
 PID=`cat $PIDFILE`
 if [[ "$PID" != "" ]]; then
-    # if [[ $GRACEFUL ]]; then
-    #     touch $KILLFILE
-    # else
+    if [[ $GRACEFUL ]]; then
+        touch $KILLFILE
+    else
         pkill -g $PID
         echo `date +%Y-%m-%dT%H-%M-%S.%6N`: Upload server killed. >> /sgm/logs/mainlog.txt
-    # fi
+    fi
     rm -f $PIDFILE
 fi
